refactor(#1367): extract tau-tools inline tests into module file
